requirements.txt
aiofiles==0.5.0
amqp==2.6.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
et-xmlfile==1.0.1
fastapi==0.60.1
future==0.18.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
numpy==1.19.0
openpyxl==3.0.4
orjson==3.3.0
pandas==1.0.5
promise==2.3
pydantic==1.6.1
pymongo==3.11.0
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
Rx==1.6.1
six==1.15.0
starlette==0.13.6
ujson==3.0.0
urllib3==1.25.9
uvicorn==0.11.6
uvloop==0.14.0
vine==1.3.0
websockets==8.1
zipp==3.1.0
